+++ src/geometry/vector.c
@@ -6,7 +6,7 @@
 /*   By: irozhkov <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/02 13:45:09 by irozhkov          #+#    #+#             */
-/*   Updated: 2025/02/14 16:24:49 by irozhkov         ###   ########.fr       */
+/*   Updated: 2025/02/17 14:25:26 by irozhkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,14 +26,14 @@ t_vector	*new_vector(double x, double y, double z)
 	return (vec);
 }
 
-float	vector_len(t_vector *vec)
+double	vector_len(t_vector *vec)
 {
 	return (sqrt((vec->x * vec->x) + (vec->y * vec->y) + (vec->z * vec->z)));
 }
 
 void	vector_normalize(t_vector *vec)
 {
-	float	len;
+	double	len;
 
 	len = vector_len(vec);
 	vec->x /= len;
