+++ src/geometry/vector_prod.c
@@ -6,13 +6,13 @@
 /*   By: irozhkov <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/05 17:37:17 by irozhkov          #+#    #+#             */
-/*   Updated: 2025/02/12 16:42:20 by irozhkov         ###   ########.fr       */
+/*   Updated: 2025/02/17 14:27:42 by irozhkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vector.h"
 
-float	vector_dot_prod(t_vector *vec1, t_vector *vec2)
+double	vector_dot_prod(t_vector *vec1, t_vector *vec2)
 {
 	return ((vec1->x * vec2->x) + (vec1->y * vec2->y) + (vec1->z * vec2->z));
 }
@@ -20,9 +20,9 @@ float	vector_dot_prod(t_vector *vec1, t_vector *vec2)
 t_vector	*vector_cross_prod(t_vector *vec1, t_vector *vec2)
 {
 	t_vector	*res;
-	float		x;
-	float		y;
-	float		z;
+	double		x;
+	double		y;
+	double		z;
 
 	x = (vec1->y * vec2->z) - (vec1->z * vec2->y);
 	y = (vec1->z * vec2->x) - (vec1->x * vec2->z);
